requirements.txt
pandas>=2.0.0
dash>=2.14.0
plotly>=5.18.0
requests>=2.28.0
gspread>=5.12.0
google-auth>=2.22.0
